Update StartPlay.py
created piece characteristics as a list of touples, linked them to their respective pieces, and created functionality to add those touples to a list, g_playBoard, every time a piece is put on the middle board. Plan to write actual win conditions tomorrow now that I have access to the piece characteristics as a list of touples. Two player will likely be done tomorrow 11/15!